Rename ws-security test to-ws-security-client test, better separate package names, test names and property names among the individual cxf test modules so that they do not clash when run together
