+++ integration-test-groups/cxf-soap/cxf-soap-ws-security-client/README.adoc
