+++ integration-test-groups/cxf-soap/cxf-soap-ws-security-client/src/main/resources/application.properties
