+++ integration-test-groups/cxf-soap/cxf-soap-ws-security-client/src/main/resources/wsdl/WssCalculatorService.wsdl
